requirements.base.txt
Django>=1.11.4,<1.12
pytz
future
django-environ>=0.4.4<0.5
djangorestframework==3.5.4
# drf-extensions==0.3.1
django-webpack-loader==0.5.0
neo4j-driver
py2neo
